+++ src/main/java/absolutelyaya/ultracraft/block/mapping/MusicListenerBlockEntity.java
@@ -87,7 +87,7 @@ public class MusicListenerBlockEntity extends AbstractListenerBlockEntity
 	}
 	
 	@Override
-	public void setAttribute(String s, String value)
+	public void setAttribute(String s, String value) throws AttributeParseException, NumberFormatException
 	{
 		switch (s)
 		{