+++ src/main/java/absolutelyaya/ultracraft/block/mapping/TimerBlockEntity.java
@@ -65,7 +65,7 @@ public class TimerBlockEntity extends AbstractTriggerBlockEntity
 	}
 	
 	@Override
-	public void setAttribute(String s, String value)
+	public void setAttribute(String s, String value) throws AttributeParseException, NumberFormatException
 	{
 		if(s.equals("start"))
 			start = Boolean.parseBoolean(value);