+++ src/main/java/absolutelyaya/ultracraft/block/mapping/TriggerBlockEntity.java
@@ -51,7 +51,7 @@ public class TriggerBlockEntity extends AbstractTriggerBlockEntity
 	}
 	
 	@Override
-	public void setAttribute(String s, String value)
+	public void setAttribute(String s, String value) throws AttributeParseException, NumberFormatException
 	{
 		switch (s)
 		{